Remove MockDependencyProvider
