+++ DuckDuckGoTests/Subscription/SubscriptionFlowViewModelTests.swift
@@ -25,7 +25,20 @@ import SubscriptionTestingUtilities
 final class SubscriptionFlowViewModelTests: XCTestCase {
     private var sut: SubscriptionFlowViewModel!
     
-    let subscriptionManager = MockDependencyProvider().subscriptionManager
+    let subscriptionManager: SubscriptionManager = {
+        let accountManager = AccountManagerMock()
+        let subscriptionService = DefaultSubscriptionEndpointService(currentServiceEnvironment: .production)
+        let authService = DefaultAuthEndpointService(currentServiceEnvironment: .production)
+        let storePurchaseManager = DefaultStorePurchaseManager()
+        return SubscriptionManagerMock(accountManager: accountManager,
+                                       subscriptionEndpointService: subscriptionService,
+                                       authEndpointService: authService,
+                                       storePurchaseManager: storePurchaseManager,
+                                       currentEnvironment: SubscriptionEnvironment(serviceEnvironment: .production,
+                                                                                   purchasePlatform: .appStore),
+                                       canPurchase: true)
+    }()
+
     let subscriptionFeatureAvailability = SubscriptionFeatureAvailabilityMock.enabled
 
     func testWhenInitWithOriginThenSubscriptionFlowPurchaseURLHasOriginSet() {
